requirements.txt
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
gunicorn==21.2.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
# pywin32==306
pyzmq==25.1.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.3.0
tornado==6.4
traitlets==5.14.2
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.1
